+++ Assets/요리시뮬레이션/script/Cooking preparation.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 618e2abf3e37e0d4e96c2385f0099ad5
+guid: ae4031c1f57695a40a0313546b616ceb
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
